split module into submodules
