+++ src/index.ts
@@ -1,20 +1,5 @@
-/**
- * Convenient type to easily represent "callable" types and to avoid using
- * "any" in client code.
- */
-export type Callable = (...args: never[]) => unknown;
-
-/**
- * Convenient type alias for constructors.
- */
-/*
- We use "any" here because this is actually a requirement on mixins at the time of this writing. Meaning,
- mixins *must* have a constructor with a single rest argument of any[]. So we thought that'd be a good default.
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type Constructor<R = object, A extends any[] = any[]> = new (
-  ...args: A
-) => R;
+export * from "./func.js";
+export * from "./guard.js";
 
 /**
  * Convenient type alias to regroup a type that can be T, null or undefined.
@@ -22,21 +7,6 @@ export type Constructor<R = object, A extends any[] = any[]> = new (
  * Semantically the opposite of {@link NonNullable}.
  */
 export type Nullable<T> = T | null | undefined;
-
-/**
- * A type alias for single element predicate functions.
- */
-export type Predicate<T> = (item: T) => boolean;
-
-/**
- * A type alias for functions that return a value of a given type without arguments.
- */
-export type Provider<T> = () => T;
-
-/**
- * A convenient type declaration for handlers used to resolve "error" type events.
- */
-export type ErrorHandler<T extends Error = Error> = (err: T) => void;
 
 /**
  * A convenience type extractor to get the inner type of an array.
@@ -74,60 +44,3 @@ export type TransitivePartial<T> = {
 export type KeyOfType<T, U> = {
   [P in keyof T]: T[P] extends U ? P : never;
 }[keyof T];
-
-/**
- * A convenient Typescript type guard function to assess that a value is a string.
- *
- * @remark
- * This function is mostly meant as a Typescript type guard. See:
- * https://www.typescriptlang.org/docs/handbook/advanced-types.html#user-defined-type-guards
- *
- * @param value - The value to test.
- * @returns Whether or not the value is a string.
- */
-export function isString(value: unknown): value is string {
-  return typeof value === "string";
-}
-
-/**
- * A convenient Typescript type guard function to assess that a value is a symbol.
- *
- * @remark
- * This function is mostly meant as a Typescript type guard. See:
- * https://www.typescriptlang.org/docs/handbook/advanced-types.html#user-defined-type-guards
- *
- * @param value - The value to test.
- * @returns Whether or not the value is a symbol.
- */
-export function isSymbol(value: unknown): value is symbol {
-  return typeof value === "symbol";
-}
-
-/**
- * A convenient Typescript type guard function to assess that a value is a number.
- *
- * @remark
- * This function is mostly meant as a Typescript type guard. See:
- * https://www.typescriptlang.org/docs/handbook/advanced-types.html#user-defined-type-guards
- *
- * @param value - The value to test.
- * @returns Whether or not the value is a number.
- */
-export function isNumber(value: unknown): value is number {
-  return typeof value === "number";
-}
-
-/**
- * A convenient Typescript type guard function to assess that a value is a function.
- *
- * @remark
- * * This function is mostly meant as a Typescript type guard. See:
- * https://www.typescriptlang.org/docs/handbook/advanced-types.html#user-defined-type-guards
- *
- * @param value - The value to test.
- * @returns Whether or not the value is a function.
- */
-// eslint-disable-next-line @typescript-eslint/ban-types
-export function isFunction(value: unknown): value is Function {
-  return typeof value === "function";
-}
